float -> double

+++ matrixFuncs/matrixMult.cpp
@@ -2,7 +2,7 @@
 #include "../linker.h"
 using namespace std;
 
-void multExt(float inputArr[5][5], int multiplier){
+void multExt(double inputArr[5][5], int multiplier){
   for(int i=0; i<5; i++){
     cout << "[";
     for(int j=0; j<5; j++){
